refactor(settings): split settings page into per-tab react-hook-form components
Break the 1515-line pages/settings.tsx monolith into a pages/settings/
folder (mirroring the models/ convention). Each tab is now its own
react-hook-form, keeping the existing persist-on-change behavior (no submit
buttons).

- Add a shared useSettingsForm hook that generalizes the original inline
  NetworkPanel pattern: dedup unchanged values, validate the field against
  the form's zod resolver, persist to PUT /settings/:key or an IPC transport,
  then patch the shared ["settings-all"] cache.
- Split tabs into recording/application/display/data/network/permissions
  panels; extract shared Row/SettingsPanel/Segment primitives and
  use-app-updater / use-permissions hooks.
- Centralize the previously-inline string unions and per-tab form schemas in
  @freestyle-voice/validations (hotkeyMode, outputMode, theme,
  audioPlaybackMode, historyRetentionPreset + *SettingsFormSchema).

Permissions has no editable fields, so it stays form-less by design.
No behavior change: dual-key audio playback write, legacy fallback chain,
IPC side-effects, custom-retention validation, and Network trim + saved-flash
are all preserved.
